Reorganize the structure of components for routing without errors
