+++ loginProject/settings.py
@@ -13,6 +13,7 @@ https://docs.djangoproject.com/en/4.2/ref/settings/
 from pathlib import Path
 from decouple import config
 import sys
+import os
 
 # Build paths inside the project like this: BASE_DIR / 'subdir'.
 BASE_DIR = Path(__file__).resolve().parent.parent
@@ -165,3 +166,5 @@ AUTH_USER_MODEL = 'loginApp.CustomUser'
 
 MAX_TITLE_LENGTH = 50
 MAX_DESCRIPTION_LENGTH = 200
+
+FRONTEND_BUILD_PATH = os.path.join(BASE_DIR, 'frontend', 'build')
